requirements.txt
pandas>=2.0.0
requests>=2.25.0
python-dotenv>=0.19.0
matplotlib>=3.5.0
numpy>=1.21.0 